Fix inner dimension offset
I noticed as part of LISIRDIII-951 that there was a wavelength offset
between data returned by ColumnarBinaryAdapter and
ColumnarBinaryAdapter2. The offset was equivalent to the number of
days since the first day in the dataset, and it turned out to be an
off-by-one error.

I recall that I subtracted one from the length because it fixed
something, but that fix must have been coincidental because removing
it fixes this issue and doesn't appear to break anything else.

See LISIRDIII-951, LATIS-610.