Fix off-by-one error on range substitution
After fixing an issue with offsets on the inner dimension, I noticed
that queries that included the end of the time dimension included an
extra time sample with a time of 0, indicating that the array
allocated to hold the results was one sample too long. (This probably
happened with the wavelengths, but it wasn't apparent because we set
the length of the inner dimension. I'm not sure.)

I think this is what inspired the original "fix" that I removed
because it cause the issue described in the previous paragraph. I
fixed the off-by-one in the wrong spot.

The real issue is that when we substituted an All range for a Bounds
range that we can actually use, we just used the length of the file
containing the particular dimension being substituted. Because the
Bounds are 0-based, we need to subtract one.

See LISIRDIII-951, LATIS-610.