Apply TimeTupleToTime within Tuples
Prior to this change, TimeTupleToTime would not look within Tuples for
Tuples named "time." Instead, Tuples not named "time" would be
returned as-is. For example:

outer: (time: (a, b, ...), ...)

Because the Tuple "time" is within the Tuple "outer", it would not be
converted by the operation.

This change makes TimeTupleToTime look within Tuples for Tuples named
"time," so Tuples named "time" nested within other Tuples will be
converted appropriately.

See NRLSSI-149.