Enable cross-building
These changes enable cross-building LaTiS between 2.11.8 (the version
we're using in most places) and whatever we set 'scalaVersion' to be.

In order to get this to work, I needed to change the versions of some
dependencies to versions that are built for both 2.11 and 2.12. The
tests still pass after these version bumps.

See LATIS-688.