Add Groovy as a runtime dependency
The logging configuration for LaTiS core requires Groovy at runtime.
By adding it as a runtime dependency here, we ensure that LaTiS core
has it available should it need it and that projects that depend on
LaTiS core don't need to remember to pull it in themselves.

There are some issues regarding local publishing that I don't
understand. Publishing Ivy-style doesn't appear to pull in Groovy, but
publishing Maven-style does.

See LATIS-692.