Make getting file size optional
This commit changes the listAllFilesWithSize methods to listAllFiles
and adds an argument that specifies whether to get the file size.

It also updates the FileListAdapter and FileListAdapterNio to use this
new method. If the TSML defines a scalar called "fileSize", the
adapter will compute the file size and include it in the dataset.
Otherwise, the file size will not be computed. (Prior to this change
it would be computed but then thrown away.)

See LATIS-575.