Add WebDirectoryListAdapter
This commit adds the WebDirectoryListAdapter, which treats the HTML
file produced when looking at the listing of a directory through a web
server as a text document and uses regular expressions (by subclassing
the RegexAdapter) to find things that match the user-specified file
name.

Where this differs from the RegexAdapter is we add the whole URL of
each file to the dataset. With the RegexAdapter alone, the best we can
do is include the whole file name, and even that required hacking in
some stuff to enable reordering of the capture groups.

See SWP-207.